+++ lizard_measure/templates/portals/organisatie-beheer.js
@@ -48,7 +48,7 @@
                 {name: 'description', title: 'beschrijving', editable: true, visible: true, width: 200, type: 'text'},
                 {name: 'group', title: 'groep', editable: true, visible: true, width: 200, type: 'text'},
                 {name: 'source', title: 'bron', editable: false, visible: true, width: 200, type: 'gridcombobox'},
-                {name: 'read_only', title: 'Alleen lezen', editable: false, visible: true, width: 50, type: 'boolean'},
+                {name: 'read_only', title: 'Alleen lezen', editable: false, visible: true, width: 50, type: 'boolean'}
            ]
         }]
 	}]